Refactor filters to some traits enable checking for implemented trait and asking column and operator configuration.
