+++ src/Filters/Concerns/HasColumn.php
@@ -1,0 +1,21 @@
+<?php
+
+
+namespace LaravelJsonApi\Eloquent\Filters\Concerns;
+
+
+trait HasColumn {
+
+    /**
+     * @var string
+     */
+    private string $column;
+
+    /**
+     * @return string
+     */
+    public function column(): string
+    {
+        return $this->column;
+    }
+}
